web_portal/backend/requirements.txt
Flask==2.3.0
Flask-SocketIO==5.3.0
python-socketio==5.8.0
eventlet==0.33.0
python-dotenv==1.0.0
discord.py==2.3.0
